Commenting clean a group of old api endpoint files and checking for complile
