+++ shifter/api/typesServerStorage.go
@@ -1,0 +1,8 @@
+package api
+
+type ServerStorage struct {
+	description string
+	storageType string
+	sourcePath  string
+	outputPath  string
+}
